move account update logic from store to client
