+++ src/client/transactions.rs
@@ -4,7 +4,7 @@ use miden_node_proto::generated::{
     requests::SubmitProvenTransactionRequest, responses::SubmitProvenTransactionResponse,
 };
 
-use crate::store::Store;
+use crate::store::{InputNoteRecord, Store};
 use miden_tx::{ProvingOptions, TransactionProver};
 use mock::procedures::prepare_word;
 use objects::{
@@ -404,7 +404,23 @@ impl Client {
             .await?;
 
         // Transaction was proven and submitted to the node correctly, persist note details and update account
-        self.store.insert_transaction_data(tx_result)?;
+        let account_id = tx_result.executed_transaction().account_id();
+        let account_delta = tx_result.account_delta();
+
+        let (mut account, seed) = self.store.get_account_by_id(account_id)?;
+
+        account
+            .apply_delta(account_delta)
+            .map_err(ClientError::AccountError)?;
+
+        let created_notes = tx_result
+            .created_notes()
+            .iter()
+            .map(|note| InputNoteRecord::from(note.clone()))
+            .collect::<Vec<_>>();
+
+        self.store
+            .insert_transaction_data(tx_result, account, seed, &created_notes)?;
 
         Ok(())
     }
